agrow-platform/py-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.3.0
requests==2.31.0
tensorflow==2.15.0
numpy==1.24.3
transformers==4.35.2
torch==2.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
email-validator==2.1.1
python-dotenv==1.0.0
